refactor: extract camera-ready preflight payload helpers
